+++ src/transformers/models/qwen4_exp/modeling_qwen4_exp.py
@@ -1700,7 +1700,7 @@ class Qwen4ExpVisionRotaryEmbedding(nn.Module):
 
         cos = self.recomposition_frequencies(cos)
         sin = self.recomposition_frequencies(sin)
-        return cos.to(dtype=x.dtype), sin.to(dtype=x.dtype)
+        return cos, sin
 
     def recomposition_frequencies(self, freq):
         """
