+++ src/transformers/models/video_llama_3/modeling_video_llama_3.py
@@ -103,7 +103,7 @@ class VideoLlama3VisionRotaryEmbedding(nn.Module):
 
         cos = self.recomposition_frequencies(cos)
         sin = self.recomposition_frequencies(sin)
-        return cos.to(dtype=x.dtype), sin.to(dtype=x.dtype)
+        return cos, sin
 
     def recomposition_frequencies(self, freq):
         """
